+++ src/components/Heading/Heading.module.css.d.ts
@@ -1,8 +1,13 @@
 declare const styles: {
+  readonly "align-center": string;
+  readonly "align-end": string;
+  readonly "align-start": string;
   readonly "color-default": string;
-  readonly "color-inverse": string;
+  readonly "color-inverted": string;
   readonly "heading": string;
+  readonly "italic": string;
   readonly "noWrap": string;
+  readonly "truncate": string;
 };
 export = styles;