+++ src/components/Heading/index.ts
@@ -3,5 +3,6 @@ export type {
   HeadingProps,
   HeadingSize,
   HeadingColor,
+  HeadingAlign,
   HeadingAs,
 } from "./Heading";