+++ src/components/Page/Page.module.css.d.ts
@@ -1,7 +1,11 @@
 declare const styles: {
+  readonly "bar": string;
+  readonly "barTitle": string;
   readonly "body": string;
   readonly "breadcrumbs": string;
+  readonly "compact": string;
   readonly "header": string;
+  readonly "large": string;
   readonly "leading": string;
   readonly "page": string;
   readonly "title": string;